ai-contact-center-solution-accelerator-main/ai-contact-center-solution-accelerator-main/eval/pyproject.toml
[project]
name = "ai-contact-centre-eval"
version = "1.0.0"
description = "End-to-end evaluation for the AI Contact Centre Solution Accelerator"
requires-python = ">=3.12"

dependencies = [
    "azure-ai-evaluation>=1.0.0",
    "azure-identity>=1.23.0",
    "openai>=1.0.0",
    "websockets>=13.0",
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eval"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.3",
]

[tool.ruff]
src = ["src"]
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["tests/**"]
